+++ src/test/java/chess/domain/position/ChessBoardTest.java
@@ -1,9 +1,11 @@
-package chess.domain;
+package chess.domain.position;
 
 import static chess.constants.TestConstants.PARAMETERIZED_TEST_NAME;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import chess.domain.ChessBoard;
+import chess.domain.GameCommand;
 import chess.domain.piece.Bishop;
 import chess.domain.piece.Color;
 import chess.domain.piece.EmptyPiece;
